Single-socket input/output nodes in node groups.
Changes the default group input/output nodes in groups to show only
a single socket. This is more convenient for keeping connections short
and put parameters closer to where they are used inside a group.

For input nodes this is purely a UI feature: they can already be
duplicated and unused sockets can be hidden (ctrl+h). The patch only
makes this the default setup of input nodes by providing all inputs
separately in the "Add Node" menu.

For output nodes the situation requires some functional changes,
because outputs have to be unique.

The "virtual" extension socket on group input/output nodes is now
disabled by default, since it adds a lot of visual noise without saving
much work (see discussion in T68733). The feature can be enabled with
a per-node flag.

Todo:
- [ ] Implement unique output handling across different node systems.

Differential Revision: https://developer.blender.org/D11853