adaptive_cloth: Mesh: split_edge_triangulate(): support for across seams
In the `Mesh` structure, `Edge` stores a tuple of `Vert`s which means,
it is based on the UV coordinates, although this is quite useful,
sometimes it is important to have the mesh operation happen across the
UV seams as well. So consider all the edges formed by the `Vert`s
stored in the `Node` of the given `Edge`'s `Vert`s. Confusing, yes,
but simple.