adaptive_cloth: Mesh: write to MeshIO: allow gaps in the Arena(s)
There can be gaps in the `Arena`s introduced due to removing elements in the middle of the `Arena`. Now this can lead to wrong indexing when writing the positions and the uvs. For this, create a `Map` between the element's arena index and it's true index in the vectors created for them.