adaptive_cloth: Mesh: better deletion of the elements
This new way of deleting elements ensures that no links can be missed out on. The deletion of the elements has to happen in a certain order leading to lesser number of errors.

This is API breaking, so all functions refering to deletion of the elements have to be updated. This includes, Mesh::split_edge_triangulate() and Mesh::collapse_edge_triangulate().