MCE: Add some options to the topbar tool settings
There are now 3 new menues on the left side of the topbar tool settings:
- Speed
- Tracking Settings
- Objects
Having the tracking speed menu now in the topbar makes more sense than
before, where it was located in the tracking settings.

The tracking presets are now located in left side of the topbar, next to
the tool settings. That way the presets are always accessible, even when
a different tool is active, which might be handy when using Detect
Features.