MCE: Trying to add mask handle props into toolbar
I tried implementing this https://developer.blender.org/P1923
To get the properties to do anything, I would need to change
add_vertex_slide to just add_vertex. add_vertex_slide is a macro, and
apparently the properties of MASK_OT_add_vertex do not have any
influence on the tool if it is set to use the macro.
But even then it only updates the handle type when activating the tool
(after switching to a different tool before). So some update thingy is
still missing.