adaptive_cloth: Mesh: get elements mutably
Useful to get the elements mutably.

This does make `Mesh` less "safe" since anyone can add or delete
elements without creating/deleting the necessary referencing between
the elements but this shouldn't be too much of a problem because it
would lead to an instant forced crash when the `Mesh` is used for
anything.