MCE: Always use mask preset for new vertex.
This is more a test, but this change makes it so that a new vertex uses
the mask tool setting for a new vertex. Not sure if this is the desired
behavior though, because before you could change the handle_type of the
active point and then every point extruded from it would use the new
handle_type.