MCE: Mask tool improvements
The mask tool now respects the settings for new handle_type in the tool
settings for new points. This makes it work more like a proper tool.

However, now it does not use the active point anymore as a reference for the handle_type
settings of new points as it was until 2.93.
So e.g. if the user starts to draw a mask with handle_type Auto and
changes the active point to Vector, every new point
drawn with the mask tool will still have handle_type Auto. Overall I
think the new behavior is better though.
Also, the user does not need to keep Ctrl pressed to draw a mask.
Instead, the tool will draw a mask point on every click, unless the mouse
is close to an existing point, then it will work like the tweak tool.

The user can hold down Ctrl to activate the add_feather_vertex_slide
tool.