Masking: Slide origin with draw_mask tool
The mask tool creates a curve point when leftclicking. When clicking
near an existing point it switches to select and move instead. The same
happens now when clicking near the origin point so it is now possible to move the
entire mask again with LMB.
Patch by Sergey, thanks a lot!