adaptive_cloth: adaptive_remesh: create an "Empty" remesh function
It is useful to call `adaptive_remesh()` outside of the cloth
simulator mainly for testing purposes. Now because the use of
templates for the function `adaptive_remesh()` and it being defined
not in the same header file (this isn't possible in this case because
of other reasons), it requires a template instantiation needs to
happen because the function is called or something like that. This
leads to problems when `adaptive_remesh()` is called from some other
file.

The easiest fix was to create a new wrapper function without templates
and call this from other files.