modifier: adaptive_remesh: UI for static remeshing of the mesh
Simple UI to allow `adaptive_remesh()` through the simple wrapper to
be called and controlled easily.

This should help with testing since the cloth simulator when applied
uses the (point) cached `Mesh` instead running the simulation which
will call `adaptive_remesh()` for us. This will need to be fixed later
but this is an easier approach to continue development the adaptive
remesher.