GPencil: Fix empty preview image
As the render preview uses a new scene, the active frame is set to 1, but grease pencil needs to have the current frame in the right place to display the strokes. Now the current frame number in the original frame is passes to scene creation function to set the same frame in the preview temporal scene.