GPencil: Add warning for non supported features
Some features as modifiers are part of the object and this asset type is datablock only and canot support them.

Now the user receives a warning about these features because they are not included in the final asset. 

To get all these features in the asset, it's necessary create Object Asset type and not Datablock type as it's done in this operator.