Geometry Nodes: Level Set Nodes (WIP)
Level sets are representations of all of the input values where a
function's output is the same. If you imagine a "position to density"
function, the corresponding level set describes a 3D surface. Disrete
Level Sets have some benefits over meshes because they always describe
a manifold shape, which makes some operations simpler. In particular,
Boolean (CSG) operations become quite straightforward. In a procedural
context, they also have the benefit of easily change-able resolution.

The OpenVDB library comes with a set of tools that getting useful
functionality out of this data structure relatively simple. This patch
adds initial versions of three level set nodes:

* **Mesh to Level Set**: Like the "Mesh to Volume" node, but creates a level set instead
** Currently the grid name is hardcoded to `level_set`
** Currently only voxel size input is available
* **Level Set Boolean**: Provides the union, difference, and intersect operations.
** The existing Boolean node should be renamed to "Mesh Boolean"
** Missing is resampling of the second input level set for a transformation.
* **Level Set Filter**: Provides "Grow", "Shrink", and "Smooth" operations.
** There are some other filters provided in the OpenVDB toolkit,
   I'm still trying to figure out their use cases though.

This patch is meant to show some possibility and give anyone interested
the chance to test. Another issue is that Blender doesn't have the ability
to display level sets at the moment, which is the idea behind this task: T88599

Differential Revision: https://developer.blender.org/D12100