Knife: Custom Knife BVH for multiple BMesh
Adds a custom Knife BVH which works with multiple BMesh's to replace the old BMBVH.
Currently, the new BVH is used in a functionally equivalent way to the old BMBVH but avoids re-allocating BVH's in multi-object edit mode.

This makes the current code slightly more complex, but will be incredibly useful when converting the knife tool code from object to world space, something that will need to be done to implement a better multi-object edit mode and further improvements to the tool.