adaptive_cloth: fix: Mesh: collapse edge: can lead to loose edges
There are some edge cases (:P) that can lead to the creation of loose
edges when collapsing the edge.

Suppose v2 has an edge that contains only 1 face and this face also
contains v1, the face was deleted but the remaining edge between v2
and ov, remained. This causes the loose edge. The simple check and
deletion of such an edge fixes this bug.