adaptive_cloth: AdaptiveMesh: sewing edge: get set of opposite edges
Given a vertex, the function is supposed to add a sewing edge to it if
possible.

Currently the function gets the set of opposite edges. An opposite
edge is an edge that is in between 2 loose edges and these loose edges
are connected to edges that connect to the given vert.

            e1   vert   e5
    e1_ov.________.________.e4_ov
         |                 |
       e2|                 |e4
         ._________________.
    e2_ov   opposite_edge   e3_ov
                 (e3)

What needs to be done: With the set of opposite edges, if the edge is
splittable then it should be split and a new edge should be added
between vert and the newly created vert (vert created when splitting
the opposite edge).