Knife: Moved code to world space
This patch migrates the knife tool code to use world space coordinates instead of object space.
This is required to implement the ability to cut across multiple objects at once and the ability to use the geometry of other objects as a cutting reference.

This introduces precision errors for objects with very large scale or location, however it may be possible to add a switch between world space and object space coordinates depending on whether the user is in multi-object edit mode or not.