Cycles X: Convert DenoiseParams to a "lite" Node
It is not supposed to be a node of a scene graph, but this change
allows to simplify serialization of the denoising parameters.

The goal is to be able to easily (de)serialize denoiser parameters
to/from image metadata, so that processing of the files can happen
outside of the scene graph (for example, on-disk tile buffer and
animation denoising).