Cycles X: Re-implement tile highlight
Since the GPUDisplay usage for offline rendering we've lost tile
highlight functionality. This change brings it back.

Introduced new RenderEngine API call to manipulate highlight state
without acquiring RenderResult (so that we don't allocate passes).

Not fully ideal since the last tile will be highlighted after render
is done until first full-frame is done, but is a cosmetic thing we
can improve in the future.

Differential Revision: https://developer.blender.org/D12519