Cycles X: make progressive multi-jitter the default sampling pattern
Now that adaptive sampling is enabled by default this was already the case in
practice, but use it now also for the non-adaptive case so the sampling patterns
stays consistent when switching that on and off.

A reason to do this now is that we will need to update all the test reference
images for the cycles-x merge, and those are all affected by the sampling pattern.