Knife: Better multi-object edit mode
The knife tool can now make cuts in multi-object edit mode which cross over multiple objects and go from the geometry of one object to multiple others.

Previously, object switching functionality was implemented but this proved insufficient to achieve the functionality contained within this patch. Instead, the dependence on a singular Object and BMEditMesh stored in the KnifeTool_OpData struct needed to be entirely removed.

This has resulted in a lot of passing around of Object pointers which could potentially be simplified using a map of some kind in the future. For now it works well.