Clip Editor: Add Clip Tools to Footage Settings
Since the old toolshelf is gone Prefetch and Set Scene Frames was only
avaible via the menu. I have now moved the old Clip Tools panel over to
the UI area in the Footage category. This is the category that is open
when a user opens a new clip, so now these opertors are quickly availble
from the sidebar for the user to set the scene frames and prefetch the
footage.
The Reload button is a bit redundant, since we have a button for that
in the footage settings panel as well, but this way the panel looks the
same as in 2.93.