Viewport Compositor: Support multiple render layers
This adds the new DRWRenderScene structure (and its sub structures) that
contains all the needed render passes for each scene present in the
compositor nodetree.

The scenes are rendered using a special option to avoid rendering overlays.

The render layer input to the GPUMaterial are now a separate structure and
a separate list of input handled by the compositor engine.

Rendering all scenes is the first thing done to avoir much trouble with

There are still issues like continuous rendering of TAA because the same
DRWData is used for all scenes.