LineArt: Better handling of boundary cases.
For lineart_LineIntersectTest2d, it fixes the output to 0 or 1 when it's
close enough, gives correct range for point_on_triangle identification.

Then it gives more information through aAligned, which will be true if
two lines completely align, this helps later identification.

Finally: it handles complex cases where one point is on the edge of
a triangle while another point is on the outside.